For property filter pages use separate files for common data and properties
