+++ pages/property-filter/hooks.page.tsx
@@ -7,7 +7,8 @@ import Button from '~components/button';
 import Box from '~components/box';
 import Header from '~components/header';
 import ScreenshotArea from '../utils/screenshot-area';
-import { allItems, columnDefinitions, TableItem, i18nStrings, filteringProperties } from './table.data';
+import { allItems, TableItem } from './table.data';
+import { columnDefinitions, i18nStrings, filteringProperties } from './common-props';
 import { useCollection } from '@cloudscape-design/collection-hooks';
 
 export default function () {
